# AchadosEAmadosPIDesktop/AchadosEAmadosPI/AchadosEamados.DTO/TipoAnimalDTO.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace AchadosEamados.DTO
{
    public class TipoAnimalDTO
    {
        public int IdTpAnimal {  get; set; }
        public string Nome {  get; set; }
    }
}