# Adotados/Adotados/AdotadosAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Achados e Amados",
    "version": "v1"
  },
  "paths": {
...
    "/api/Animal/{id}": {
      "get": {
        "tags": [
          "Animal"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AnimalDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Animal"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AnimalDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Animal"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Funcionario": {
      "get": {
        "tags": [
          "Funcionario"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/FuncionarioDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Funcionario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FuncionarioDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/FuncionarioDTO"
            }
          }
        }
      }
    },
...
    "/api/Funcionario/{id}": {
      "get": {
        "tags": [
          "Funcionario"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/FuncionarioDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Funcionario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FuncionarioDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Funcionario"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/TipoAnimal": {
      "get": {
        "tags": [
          "TipoAnimal"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TipoAnimalDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TipoAnimal"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TipoAnimalDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TipoAnimalDTO"
            }
          }
        }
      }
    },
    "/api/TipoAnimal/{id}": {
      "get": {
        "tags": [
          "TipoAnimal"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TipoAnimalDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "TipoAnimal"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TipoAnimalDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "TipoAnimal"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/TipoFuncionario": {
      "get": {
        "tags": [
          "TipoFuncionario"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TipoFuncionarioDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TipoFuncionario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TipoFuncionarioDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TipoFuncionarioDTO"
            }
          }
        }
      }
    },
    "/api/TipoFuncionario/{id}": {
      "get": {
        "tags": [
          "TipoFuncionario"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TipoFuncionarioDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "TipoFuncionario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TipoFuncionarioDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "TipoFuncionario"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AnimalDTO": {
      "type": "object",
      "properties": {
        "idAnimal": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "idade": {
          "type": "string"
        },
        "ehVacinado": {
          "type": "boolean"
        },
        "urlImgAnimal": {
          "type": "string"
        },
        "descricao": {
          "type": "string"
        },
        "dtaCadastro": {
          "format": "date-time",
          "type": "string"
        },
        "idTpFuncionario": {
          "format": "int32",
          "type": "integer"
        },
        "idTpAnimal": {
          "format": "int32",
          "type": "integer"
        },
        "sexo": {
          "type": "string"
        },
        "raca": {
          "type": "string"
        },
        "ehCastrado": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "FuncionarioDTO": {
      "type": "object",
      "properties": {
        "idFuncionario": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "senha": {
          "type": "string"
        },
        "dtaNascimento": {
          "format": "date-time",
          "type": "string"
        },
        "cpf": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        },
        "sexo": {
          "type": "string"
        },
        "urlImgFuncionario": {
          "type": "string"
        },
        "idTpFuncionario": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "TipoAnimalDTO": {
      "type": "object",
      "properties": {
        "idTpAnimal": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "TipoFuncionarioDTO": {
      "type": "object",
      "properties": {
        "idTpFuncionario": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira o JWT no formato: Bearer {seu_token}"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}